+++ Common/src/main/java/com/hypherionmc/sdlink/core/config/SDLinkCompatConfig.java
@@ -4,10 +4,7 @@ import com.hypherionmc.craterlib.core.config.AbstractConfig;
 import com.hypherionmc.craterlib.core.config.ConfigController;
 import com.hypherionmc.craterlib.core.config.annotations.NoConfigScreen;
 import com.hypherionmc.craterlib.core.config.formats.TomlConfigFormat;
-import com.hypherionmc.sdlink.core.config.impl.compat.CommonCompat;
-import com.hypherionmc.sdlink.core.config.impl.compat.MaintenanceModeCompat;
-import com.hypherionmc.sdlink.core.config.impl.compat.PlayerReviveCompat;
-import com.hypherionmc.sdlink.core.config.impl.compat.RoleSyncCompat;
+import com.hypherionmc.sdlink.core.config.impl.compat.*;
 import com.hypherionmc.sdlink.core.discord.BotController;
 import org.apache.commons.io.FileUtils;
 import shadow.hypherionmc.moonconfig.core.conversion.ObjectConverter;
@@ -24,7 +21,7 @@ public final class SDLinkCompatConfig extends AbstractConfig<SDLinkCompatConfig>
     // DO NOT REMOVE TRANSIENT HERE... OTHERWISE, THE STUPID CONFIG LIBRARY
     // WILL TRY TO WRITE THESE TO THE CONFIG
     public transient static SDLinkCompatConfig INSTANCE;
-    public transient static int configVer = 3;
+    public transient static int configVer = 4;
     public transient static boolean hasConfigLoaded = false;
     public transient static boolean wasReload = false;
 
@@ -51,6 +48,9 @@ public final class SDLinkCompatConfig extends AbstractConfig<SDLinkCompatConfig>
     @Path("ftbranks")
     @SpecComment("FTB Ranks Rank syncing")
     public RoleSyncCompat ftbRanksCompat = new RoleSyncCompat();
+
+    @Path("vanish")
+    public VanishCompat vanishCompat = new VanishCompat();
 
     public SDLinkCompatConfig() {
         this(false);